% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONCERT-package.R
\docType{package}
\name{CONCERT-package}
\alias{CONCERT}
\alias{CONCERT-package}
\title{CONCERT: What the Package Does (One Line, Title Case)}
\description{
This package implements the Variational Spike-and-Slab Transfer Regression method, providing tools for efficient statistical analysis.
}
\author{
\strong{Maintainer}: Bao Jiayi \email{baojiayi689@gmail.com}

}
\keyword{internal}
