% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concert-binary.R
\name{eval_pre}
\alias{eval_pre}
\title{Evaluate prediction error}
\usage{
eval_pre(est, X_test = NULL, y_test = NULL)
}
\arguments{
\item{est}{A vector representing the estimated coefficients.}

\item{X_test}{A matrix of test feature data.}

\item{y_test}{A vector of true labels for the test data.}
}
\value{
A numeric value representing the prediction error.
}
\description{
This function calculates the prediction error for binary classification based on logistic regression.
}
\examples{
est <- c(0.5, -0.2)
X_test <- matrix(c(1, 2, 3, 4), ncol=2)
y_test <- c(1, 0)
eval_pre(est, X_test, y_test)
}
