% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concert-binary.R
\name{perf}
\alias{perf}
\title{Performance metrics for regression}
\usage{
perf(theta, theta0)
}
\arguments{
\item{theta}{A vector of estimated coefficients.}

\item{theta0}{A vector of true coefficients.}
}
\value{
A vector of two values: TP and FP rates.
}
\description{
This function calculates True Positive (TP) and False Positive (FP) rates for the regression model.
}
\examples{
theta <- c(1, 0, 1, 0, 0)
theta0 <- c(1, 0, 0, 1, 0)
perf(theta, theta0)
}
