% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concert-binary.R
\name{sample_ind}
\alias{sample_ind}
\title{Create indices for multiple source datasets}
\usage{
sample_ind(n_vec, num_K)
}
\arguments{
\item{n_vec}{A vector of integers indicating the number of data points in each dataset.}

\item{num_K}{An integer indicating the number of datasets (excluding the target dataset).}
}
\value{
A list of length \code{num_K}, where each element is a vector of indices corresponding to the rows of the input data for that dataset.
}
\description{
This function generates a list of indices to split data into multiple source datasets based on the provided vector of dataset sizes.
}
\examples{
n_vec <- c(100, 150, 200)
num_K <- 2
sample_ind(n_vec, num_K)
}
