% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concert-binary.R
\name{Solo_R}
\alias{Solo_R}
\title{Variational Spike-and-Slab Regression (Solo Model)}
\usage{
Solo_R(X, y, threshold = 1e-06, max_iter = 1000, eta = 1, q_0 = 0.05)
}
\arguments{
\item{X}{A matrix of size \verb{n x p} representing the feature data, where \code{n} is the number of observations and \code{p} is the number of features.}

\item{y}{A vector of length \code{n} representing the binary target variable for the regression model.}

\item{threshold}{A numeric value specifying the convergence threshold. The algorithm stops if the difference between successive estimates of the coefficients is smaller than this value. Default is \code{1e-6}.}

\item{max_iter}{An integer specifying the maximum number of iterations. Default is \code{1000}.}

\item{eta}{A numeric value representing the learning rate or scale factor for the variance of the coefficients. Default is \code{1}.}

\item{q_0}{A numeric value representing the prior probability of a coefficient being zero (the "spike" in the spike-and-slab prior). Default is \code{0.05}.}
}
\value{
A list containing the following elements:
\itemize{
\item \code{m_beta0}: A vector of the estimated regression coefficients for the target model.
\item \code{gamma_Z}: A vector of the posterior probabilities that each coefficient is non-zero (indicator variable for each coefficient).
\item \code{V_beta0}: A vector of the variances of the estimated coefficients.
}
}
\description{
This function implements a Variational Spike-and-Slab Regression model, which is used to estimate the regression coefficients for a target dataset using logistic regression and a variational inference approach.
}
\examples{
set.seed(123)
X <- matrix(rnorm(200), ncol=2)
y <- sample(0:1, 100, replace=TRUE)
result <- Solo_R(X, y, threshold=1e-6, max_iter=1000, eta=1, q_0=0.05)
result$m_beta0
result$gamma_Z
result$V_beta0

}
